Refactored logging into multiple files
