+++ LogRawBinary.cpp
@@ -1,0 +1,13 @@
+#include "LogRawBinary.h"
+
+LogRawBinary::LogRawBinary(char* basename) {
+  char filename[256];
+  snprintf(filename,sizeof(filename)-1,"%s/%s",recordPath,basename);
+  stream=fopen(filename,"w");
+  setbuf(stream,nullptr); //Turn off buffering for this file
+}
+
+LogRawBinary::~LogRawBinary() {
+  fclose(stream);
+}
+
